backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.28.1
python-dotenv==1.2.1
asyncpg==0.30.0
bcrypt==4.3.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
email-validator==2.2.0
better-profanity==0.7.0
slowapi==0.1.9